+++ tests/pose_optimizer/tools.py
@@ -5,10 +5,6 @@ from modules.pose_optimizer.optimizer import Optimizer, Keyframe
 from tests.tools.data_manager import DataManager
 
 
-def get_random():
-    return np.random.rand(1)[0]
-
-
 def setup(
     feature_points_position,
     cameras_position,
@@ -65,34 +61,3 @@ def logging_data(optimizer, data_manager):
 
     data_manager.logging_trajectory(fp_data, kf_data)
 
-
-def get_random_points(xy_scale, z_scale, z_offset, num):
-    points = [
-        np.array(
-            [
-                (get_random() - 0.5) * xy_scale,
-                (get_random() - 0.5) * xy_scale,
-                get_random() * z_scale + z_offset,
-            ]
-        )
-        for i in range(num)
-    ]
-    return points
-
-
-def get_random_rotation(offset, scale, num):
-    random_vector = [np.random.normal(offset, scale, 3) for p in range(num)]
-    return [scipy_R.from_euler("xyz", v).as_matrix() for v in random_vector]
-
-
-def get_square_position(x, y, z):
-    points = [[x, y, z], [x, -y, z], [-x, y, z], [-x, -y, z]]
-    return [np.array(p) for p in points]
-
-
-def get_position(points):
-    return [np.array(p) for p in points]
-
-
-def get_rotation(vectors):
-    return [scipy_R.from_euler("xyz", v).as_matrix() for v in vectors]
