+++ tests/tools/results_analyzer.py
@@ -2,10 +2,12 @@ import pandas as pd
 import pathlib
 import argparse
 import sys
-from tests.tools.fig_maker import FigMaker
+from mpl_toolkits.mplot3d import Axes3D
 
 current_dir = pathlib.Path(__file__).resolve().parent
 sys.path.append(str(current_dir) + "/../../")
+
+from tests.tools.fig_maker import FigMaker
 
 ap = argparse.ArgumentParser()
 ap.add_argument("log_file_name")
